% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cjsg.R
\name{cjsg}
\alias{cjsg}
\title{Funcao que faz o download da lista de processos na Consulta de Julgados do Segundo Grau (CJSG)
do TJSP.}
\usage{
cjsg(livre = NULL, ementa = NULL, relator = NULL, prolator = NULL,
  classes = NULL, assuntos = NULL, comarcas = NULL, orgaos = NULL,
  data_inicial = NULL, data_final = NULL, origem = c("2grau",
  "colegio_recursal"), tipo = c("acordao", "acordo", "monocratica"),
  min_pag = 1, max_pag = 10, salvar = FALSE, path = "")
}
\description{
Funcao que faz o download da lista de processos na Consulta de Julgados do Segundo Grau (CJSG)
do TJSP.
}

