% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpo-pg.R
\name{cpo_pg}
\alias{cpo_pg}
\title{Funcao que faz o download das informacoes de um processo de primeiro
grau (PG) no TJSP.}
\usage{
cpo_pg(processos, path = NULL)
}
\description{
Retorna um data.frame com os metadados basicos e andamentos do processo
}

