% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpo-pg.R
\name{cpo_pg_um}
\alias{cpo_pg_um}
\title{Funcao que faz o download das informacoes de um processo de primeiro
grau (PG) no TJSP.}
\usage{
cpo_pg_um(p, path)
}
\description{
Funcao que faz o download das informacoes de um processo de primeiro
grau (PG) no TJSP.
}

