% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpo-sg.R
\name{cpo_sg}
\alias{cpo_sg}
\title{Funcao que faz o download das informacoes de um processo de segundo
grau (SG) no TJSP.}
\usage{
cpo_sg(processos, path = NULL)
}
\description{
Retorna um data.frame com os metadados basicos e andamentos do processo
}

