% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpo-sg.R
\name{cpo_sg_um}
\alias{cpo_sg_um}
\title{Funcao que faz o download das informacoes de um processo de segundo
grau (SG) no TJSP.}
\usage{
cpo_sg_um(p, path)
}
\description{
Funcao que faz o download das informacoes de um processo de segundo
grau (SG) no TJSP.
}

