% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-pecas.R
\name{download_arqs}
\alias{download_arqs}
\title{Funcao que faz o download de todos os arquivos}
\usage{
download_arqs(processos, login, senha, path)
}
\description{
Funcao que faz o download de todos os arquivos
}

