% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/download-pecas.R
\name{download_peticoes}
\alias{download_peticoes}
\title{Funcao que faz o download das peticoes.}
\usage{
download_peticoes(processos, login, senha, path)
}
\description{
Funcao que faz o download das peticoes.
}

