% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_cjpg.R
\name{list_cjpg}
\alias{list_cjpg}
\title{Funcao que faz o download da lista de processos na Consulta de Julgados do Primeiro Grau (CJPG)
do TJSP.}
\usage{
list_cjpg(livre = NULL, classes = NULL, assuntos = NULL,
  data_inicial = NULL, data_final = NULL, varas = NULL, min_pag = 1,
  max_pag = 10, salvar = FALSE, path = "")
}
\description{
Funcao que faz o download da lista de processos na Consulta de Julgados do Primeiro Grau (CJPG)
do TJSP.
}

