% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_cjpg.R
\name{n_cjpg}
\alias{n_cjpg}
\title{Calcula quantas paginas tem uma query no CJPG}
\usage{
n_cjpg(livre = NULL, classes = NULL, assuntos = NULL,
  data_inicial = NULL, data_final = NULL, varas = NULL)
}
\description{
Calcula quantas paginas tem uma query no CJPG
}

