% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cjsg.R
\name{n_cjsg}
\alias{n_cjsg}
\title{Calcula quantas paginas tem uma query no CJPG}
\usage{
n_cjsg(livre = NULL, ementa = NULL, relator = NULL, prolator = NULL,
  classes = NULL, assuntos = NULL, comarcas = NULL, orgaos = NULL,
  data_inicial = NULL, data_final = NULL, origem = c("2grau",
  "colegio_recursal"), tipo = c("acordao", "acordo", "monocratica"))
}
\description{
Calcula quantas paginas tem uma query no CJPG
}

